SIAD link speed plugin
Closes: VRVDR-50258

The SIAD-specific must statement introduced to deal with interface breakout on dp0xe20-23 and separately dp0xe24-27 added by VRVDR-49705 has introduced a significant time penalty for configuration validation on the SIAD.  The must statement relies on a certain amount of 'brute force' and is costly in terms of processing time.  The simplest fix is to write a custom function that can replace it on the SIAD using the configd:must statement in YANG.

This commit adds the plugin - a separate PR will make the necessary yang changes.

I've tested on a SIAD (ed-siad-8) and the time taken for YANG validation reduces from 3s to 1s, so a very noticeable saving, with the reference configuration we use to test performance.